backend/requirements.txt
Django~=5.2.1 
psycopg2-binary # For PostgreSQL connectivity
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
requests==2.32.3
drf-yasg
python-dateutil
pytest
pytest-django
factory_boy
pytest-mock
